Update: Modularise PageTitle component
